analysing sph
