+++ avaframe/com1DFAPy/timeDiscretizations.py
@@ -5,17 +5,11 @@
 """
 
 # Load modules
-import os
-import glob
 import logging
 import numpy as np
 
 # Local imports
-import avaframe.in3Utils.fileHandlerUtils as fU
-import avaframe.in2Trans.ascUtils as IOf
-from avaframe.in3Utils import cfgUtils
-from avaframe.in3Utils import logUtils
-import avaframe.com1DFAPy.com1DFAtools as DFAtls
+import avaframe.com1DFAPy.DFAtools as DFAtls
 
 
 # create local logger
